refactor: POST 요청의 Request body 처리를 Request 객체로 이동
